+++ CHANGE_LOG.md
@@ -4,7 +4,8 @@
 
 * [fix] Minor tweaks
 * [feature] Improve preference window's overall design
-* [feature, experimental] Add snippet feature, window
+* [feature] Change sidebar's icons
+* [feature, experimental] Add Snippet feature, window, preference page
 * [feature, experimental] Add Universal action window, preference page
 
 # 0.10.8